+++ src/Interfaces/Standalone/hcoi_standalone_mod.F90
@@ -2368,20 +2368,6 @@ CONTAINS
        ENDIF
     ENDIF
 
-    !%%%%% PBL height  %%%%%
-    IF ( ExtState%PBLH%DoUse ) THEN
-       Name = 'PBLH'
-       CALL ExtDat_Set( HcoState,     ExtState%PBLH,                         &
-                        TRIM( Name ), RC,                   FIRST=FIRST     )
-       IF ( RC /= HCO_SUCCESS ) THEN
-          ErrMsg = 'Could not find quantity "' // TRIM( Name )            // &
-                   '" for the HEMCO standalone simulation!'
-          CALL HCO_Error( ErrMsg, RC, ThisLoc )
-          CALL HCO_Leave( HcoState%Config%Err, RC )
-          RETURN
-       ENDIF
-    ENDIF
-
     !%%%%% Sensible heat flux %%%%%
     IF ( ExtState%HFLUX%DoUse ) THEN
        Name = 'HFLUX'